Add ability to return results in a nested format
- Move fetchDataForAllYears to separate file
